+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.22.0] - 2025-05-19
+### Added
+- Default "Transfers" category group with Adjustment, Goal Contribution and Loan Payment categories.
+- New transactions default to the Adjustment category when Transfer is selected.
+
 ## [0.21.0] - 2025-05-19
 ### Added
 - Progress bar for category budgets.